+++ packages/aws/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @authhero/aws-adapter
 
+## 0.15.0
+
+### Minor Changes
+
+- a8e70e6: Update schemas to remove old fallbacks
+
+### Patch Changes
+
+- Updated dependencies [a8e70e6]
+  - @authhero/adapter-interfaces@0.129.0
+
 ## 0.14.0
 
 ### Minor Changes