+++ packages/cloudflare/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @authhero/cloudflare-adapter
 
+## 2.18.5
+
+### Patch Changes
+
+- Updated dependencies [3b4445f]
+  - @authhero/kysely-adapter@10.92.0
+
+## 2.18.4
+
+### Patch Changes
+
+- Updated dependencies [a8e70e6]
+  - @authhero/adapter-interfaces@0.129.0
+  - @authhero/kysely-adapter@10.91.0
+
 ## 2.18.3
 
 ### Patch Changes